LayoutTests/fast/scrolling/mac/event-region-prevent-default-with-sublayer-expected.txt

Test scroll over composited layer
Got wheel event 10
Got wheel event 100
PASS windowScrollEventCount is 0
PASS successfullyParsed is true

TEST COMPLETE
